+++ src/ui/tabs/overlay.py
@@ -55,14 +55,14 @@ class OverlayTab(TabBase):
 
     def _build_overlay_switch(self, card) -> None:
         right = self._setting_row(card, "Show Overlay", "Display transcription text while dictating")
-        self._overlay_var = ctk.BooleanVar(value=self._settings.overlay_enabled)
+        self._overlay_var = self._bool_var(value=self._settings.overlay_enabled)
         themed_switch(
             right, self._colors, variable=self._overlay_var,
             command=self._on_overlay_toggle,
         ).pack()
 
         right = self._setting_row(card, "Audio Level Indicator", "Small line that moves with your voice")
-        self._level_var = ctk.BooleanVar(value=self._settings.overlay_show_level)
+        self._level_var = self._bool_var(value=self._settings.overlay_show_level)
         themed_switch(
             right, self._colors, variable=self._level_var,
             command=self._on_level_toggle,
@@ -76,7 +76,7 @@ class OverlayTab(TabBase):
             "bottom-left": "Bottom Left", "bottom-right": "Bottom Right",
         }
         current_pos = pos_labels.get(self._settings.overlay_position, "Bottom Center")
-        self._pos_var = ctk.StringVar(value=current_pos)
+        self._pos_var = self._string_var(value=current_pos)
         self._dropdown(
             right, variable=self._pos_var,
             values=list(pos_labels.values()),